+++ Gulpfile.js
@@ -40,14 +40,16 @@ toolkit.extendConfig(
 			zip: [
 				'./**/*',
 				'!./*.zip',
+				'!./git',
+				'!./git/**/*',
 				'!./node_modules',
 				'!./node_modules/**/*',
 				'!./vendor',
 				'!./vendor/**/*',
-				'!./git',
-				'!./git/**/*',
-				'!./lib/.git',
-				'!./lib/.git/**/*'
+				'./vendor/composer/*.php',
+				'./vendor/seothemes/child-theme-library/src/*.php',
+				'./vendor/tgmpa/tgm-plugin-activation/languages/*',
+				'./vendor/tgmpa/tgm-plugin-activation/class-tgm-plugin-activation.php'
 			]
 		},
 		css: {